Break large tests.py files into smaller files
